// web/src/api/generated/models/PrivateCollectionItemDetail.ts
export interface PrivateCollectionItemDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof PrivateCollectionItemDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionItemDetail
     */
    collectionId: string;
    /**
     * 
     * @type {boolean}
     * @memberof PrivateCollectionItemDetail
     */
    completed: boolean;
    /**
     * 
     * @type {Date}
     * @memberof PrivateCollectionItemDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionItemDetail
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof PrivateCollectionItemDetail
     */
    position: number;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionItemDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof PrivateCollectionItemDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof PrivateCollectionItemDetail
     */
    version: number;
}